fix: dropping item from GUI causing crystal explosion The ANIMATION packet for WINDOW_CLICK (dropping from GUI) comes before the WINDOW_CLICK packet, meaning lastPacket is useless for this check
To get around this, introduced a new field "ignoreAnim" that ignores the current ANIMATION packet if this is true. Note that this is checked in the main thread in AnimationListener, meaning this field was most likely already updated by the time the main task runs.

Also, LastPacketListener no longer uses the main thread. Instead, it updates the field instantly but consequently the AnimationListener must save the last packet prior to running the main task.